Orchestrator and notificaiton service communication completed

+++ src/main/java/com/example/notificationservice/DTOs/MoveData.java
@@ -1,10 +1,12 @@
 package com.example.notificationservice.DTOs;
 
+import java.time.Instant;
+
 public record MoveData(
         Integer moveNumber,
         String movePlayedBy,
         String playerColor,
-        Integer evalBefore,
-        Integer evalAfter,
+        Double evalBefore,
+        Double evalAfter,
         Integer timeRemaining
 ){}
